pyproject.toml
[build-system]
requires = ['hatchling']
build-backend = 'hatchling.build'

[project]
name = "congress"
version = "0.1"
authors = [
	{ name="Tom Casavant", email="tfcasavant@gmail.com" }
]
description = "A wrapper for the congress.gov API"
readme = "README.md"
dependencies = [
	"requests",
	"python-dateutil"
]
requires-python = ">=3.8"
classifiers = [
	"Programming Language :: Python :: 3",
	"License :: OSI Approved :: MIT License",
	"Operating System :: OS Independent",
	"Topic :: Software Development :: Libraries :: Python Modules",
	"Topic :: Utilities",
	"Topic :: Internet"
]
keywords = ["congress"]

[project.urls]
Homepage = ""
Issues = ""

[tool.hatch.build.targets.wheel]
packages = ["src/congress"]
